pyproject.toml
[project]
name = "tick-listo-cli"
version = "0.1.0"
description = "Tick Listo CLI - Your Ultimate Task Management - Where Productivity Meets Elegance"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "rich>=14.3.1",
    "python-dateutil>=2.8.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
